refactor(chess): finish Spring migration
- remove unused files and methods
