+++ src/main/java/chess/controller/HomeController.java
@@ -1,31 +1,34 @@
 package chess.controller;
 
-import static chess.web.WebUtils.render;
-import static spark.Spark.get;
-import static spark.Spark.post;
-
-import chess.dto.CreateGameDto;
+import chess.dto.GameCountDto;
 import chess.dto.SuccessResponseDto;
 import chess.service.ChessService;
-import chess.web.WebUtils;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.ModelAndView;
 
+@RestController
+@RequestMapping("/")
 public class HomeController {
 
-    private static final String HOME_ROUTE = "/";
-    private static final String HTML_TEMPLATE_PATH = "home.html";
+    private static final String HTML_TEMPLATE_PATH = "home";
+    private static final String HOME_DYNAMIC_PROPERTIES = "gameCount";
 
     private final ChessService chessService = ChessService.getInstance();
 
-    public void initRouteHandler() {
-        get(HOME_ROUTE, (req, res) -> render(chessService.countGames(), HTML_TEMPLATE_PATH));
-        post(HOME_ROUTE, (req, res) -> {
-            res.type(WebUtils.JSON_CONTENT_TYPE);
-            return new SuccessResponseDto(getGameCreatedResponse()).toJson();
-        });
+    @GetMapping
+    public ModelAndView home() {
+        ModelAndView modelAndView = new ModelAndView();
+        modelAndView.setViewName(HTML_TEMPLATE_PATH);
+        GameCountDto gameCountDto = chessService.countGames();
+        modelAndView.addObject(HOME_DYNAMIC_PROPERTIES, gameCountDto);
+        return modelAndView;
     }
 
-    private String getGameCreatedResponse() {
-        CreateGameDto gameCreated = chessService.initGame();
-        return gameCreated.toJson();
+    @PostMapping
+    public SuccessResponseDto createGame() {
+        return new SuccessResponseDto(chessService.initGame());
     }
 }
