--- src/main/resources/templates/index.hbs
@@ -1,9 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <title>모두의 체스</title>
-</head>
-<body>
-<h1>Hello World</h1>
-</body>
-</html>
